Add predirectional to Geocodio::Address
This adds the previously missed support for predirectionals to instances
of Geocodio::Address. When given an address such as '54 West Colorado
Boulevard Pasadena CA 91105', the result contains a predirectional of
'W'. Previously, this predirectional ended up being dropped as an
attribute (though it remained in the formatted_address/to_s method,
which is the main interface used).

Signed-off-by: David Celis <me@davidcel.is>