Return empty address sets for bad batch geocodes
When geocoding a single address, raising a `Geocodio::Client::Error` is
expected if the address was not parseable by Geocod.io. However, during
batch geocoding, it is possible that only some addresses were not
parseable. For those addresses, Geocod.io returns a hash with an error
message but will return good results for all other addresses. In this
case, we don't want to raise an error. Instead, simply return an empty
`Geocodio::AddressSet`

Fixes #3

Signed-off-by: David Celis <me@davidcel.is>